pyproject.toml
[tool.poetry]
name = "kden-stock"
version = "0.1.0"
description = ""
authors = ["ghostrasiel <ghostrasiel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pandas = "^2.1.3"
google-cloud-storage = "1.42.0"
google-cloud-pubsub = "2.18.4"
pyarrow = "14.0.1"
numpy = "1.26.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
